Update tableData for this replacement type change
